Prevent crash when TXPower is set to 5+
These changes are required to fix the coordinator crashing when the TX power is set to 5+.
In SDK 7.10.02.23 TI made changes to `macSetTxPowerVal` which introduced this issue.
The patch reverts these changes.
The patched `maclib_*.a` files were provided by a TI employee and will only work for the 7.41.00.17 SDK.