sdk: packdev_t: Increment pointer address, not pointer value
If pDev is not NULL, everything in the buffer from linkInfo.InFrmCntr
onwards was also garbage:
The *pBuf+=4 is definitely a bug, it increments the value, not the pointer
address. So linkInfo.InFrmCntr is written to the buffer, then
linkInfo.InFrmCntr in the buffer is incremented by 4, and then the first
two bytes (because the pointer was not incremented) are overwritten with
linkInfo.TxFailure. I replaced it with the correct pBuf +=4.

See https://github.com/Koenkk/zigbee2mqtt/issues/13478#issuecomment-1501085509

@slugzero

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>