mt: Introduce LED API
Allow the LED to be turned on/off via the API with a new command,
MT_UTIL_LED_CONTROL (0x0A).

The led is setup as 'CONFIG_LED_GREEN' on many boards and typically is
controlled by IO7. Some boards may not populate these components or not
even route them, but this is not an issue, as this is made available (or
not) via the board/sysconfig.

@koenkk

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>