NPI_UART: Use DisplayUart2 configuration instead of hardcoding
The NPI configuration is for the most part hardcoded. While it is
possible to set a define from the commandline it is not ideal. Instead,
lets use the DisplayUart2 configuration as configured in sysconfig.

Hardcoding this as such should be acceptable, as the
`CONFIG_DISPLAY_UART` is already used as index which is a hard
requirement to using the Display component, and currently we are
just hardcoding some values anyway.

The 'old' behavior will remain in effect for the baud rate. Changing of
the baud rate is only triggered if the configuration parameter
`-DNPI_UART_BR=0` is supplied. Thus someone explicitly has to set the
baudrate to zero via said define to trigger this behavior.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>